+++ rollup.config.js
@@ -33,23 +33,11 @@ const createTestEntries = () => {
 		return [];
 	}
 
-	const mainSpecs = jetpack.find('src/main', { matching: '*.spec.js' });
 	const rendererSpecs = [
 		...jetpack.find('src/scripts', { matching: '*.spec.js' }),
 		...jetpack.find('src/preload', { matching: '*.spec.js' }),
 	];
 	return [
-		...mainSpecs.length
-			? [{
-				input: mainSpecs,
-				...bundleOptions,
-				output: {
-					dir: 'app/main.specs',
-					format: 'cjs',
-					sourcemap: true,
-				},
-			}]
-			: [],
 		...rendererSpecs.length
 			? [{
 				input: rendererSpecs,
@@ -105,18 +93,7 @@ export default [
 			{
 				file: 'app/preload.js',
 				format: 'cjs',
-				sourcemap: true,
-			},
-		],
-	},
-	{
-		input: 'src/preload.js',
-		...bundleOptions,
-		output: [
-			{
-				file: 'app/preload.js',
-				format: 'cjs',
-				sourcemap: true,
+				sourcemap: 'inline',
 			},
 		],
 	},