+++ src/app.js
@@ -1,6 +1,11 @@
-import { start } from './scripts/start';
 import { setupErrorHandling } from './errorHandling';
+import { setupI18next } from './i18n';
+import attachEvents from './scripts/events';
 
+const initialize = async () => {
+	setupErrorHandling('renderer');
+	await setupI18next();
+	await attachEvents();
+};
 
-setupErrorHandling('renderer');
-start();
+initialize();