+++ src/main.js
@@ -1,40 +1,41 @@
-import { app } from 'electron';
+import path from 'path';
+
+import { app, BrowserWindow } from 'electron';
 import setupElectronReload from 'electron-reload';
+import jetpack from 'fs-jetpack';
 
 import { setupErrorHandling } from './errorHandling';
-import appData from './main/appData';
-import './main/basicAuth';
-import { processDeepLink } from './main/deepLinks';
-import './main/updates';
-import { getMainWindow } from './main/mainWindow';
-import './main/dialogs/about';
-import './main/dialogs/screenshare';
-import './main/dialogs/update';
-import i18n from './i18n';
 
-export { default as dock } from './main/dock';
-export { default as menus } from './main/menus';
-export { default as tray } from './main/tray';
-export { default as certificate } from './main/certificateStore';
-
-if (process.env.NODE_ENV) {
+if (process.env.NODE_ENV === 'development') {
 	setupElectronReload(__dirname, {
 		electron: process.execPath,
 	});
 }
 
-async function prepareApp() {
+const preventEvent = (event) => event.preventDefault();
+
+const prepareApp = () => {
 	setupErrorHandling('main');
 
 	app.setAsDefaultProtocolClient('rocketchat');
 	app.setAppUserModelId('chat.rocket');
 
-	await appData.initialize();
+	const dirName = process.env.NODE_ENV === 'production' ? app.name : `${ app.name } (${ process.env.NODE_ENV })`;
+
+	app.setPath('userData', path.join(app.getPath('appData'), dirName));
+
+	if (process.argv[2] === '--reset-app-data') {
+		const dataDir = app.getPath('userData');
+		jetpack.remove(dataDir);
+		app.relaunch({ args: [process.argv[1]] });
+		app.exit();
+		return;
+	}
 
 	const canStart = process.mas || app.requestSingleInstanceLock();
 
 	if (!canStart) {
-		app.quit();
+		app.exit();
 		return;
 	}
 
@@ -45,25 +46,41 @@ async function prepareApp() {
 		app.disableHardwareAcceleration();
 	}
 
+	app.on('certificate-error', preventEvent);
+	app.on('login', preventEvent);
+	app.on('open-url', preventEvent);
 	app.on('window-all-closed', () => {
 		app.quit();
 	});
+};
 
-	app.on('open-url', (event, url) => {
-		event.preventDefault();
-		processDeepLink(url);
+const createMainWindow = () => {
+	const mainWindow = new BrowserWindow({
+		width: 1000,
+		height: 600,
+		minWidth: 600,
+		minHeight: 400,
+		titleBarStyle: 'hidden',
+		show: false,
+		webPreferences: {
+			webviewTag: true,
+			nodeIntegration: true,
+		},
 	});
 
-	app.on('second-instance', (event, argv) => {
-		argv.slice(2).forEach(processDeepLink);
-	});
-}
+	mainWindow.on('close', preventEvent);
 
-(async () => {
-	await prepareApp();
+	mainWindow.webContents.on('will-attach-webview', (event, webPreferences) => {
+		delete webPreferences.enableBlinkFeatures;
+	});
+
+	mainWindow.loadFile(`${ app.getAppPath() }/app/public/app.html`);
+};
+
+const initialize = async () => {
+	prepareApp();
 	await app.whenReady();
-	await i18n.initialize();
-	app.emit('start');
-	await getMainWindow();
-	process.argv.slice(2).forEach(processDeepLink);
-})();
+	createMainWindow();
+};
+
+initialize();