+++ src/preload.js
@@ -1,7 +1,4 @@
-import { ipcRenderer } from 'electron';
-
 import { setupErrorHandling } from './errorHandling';
-import i18n from './i18n';
 import setupContextMenuPreload from './preload/contextMenu';
 import setupEventsPreload from './preload/events';
 import setupJitsiPreload from './preload/jitsi';
@@ -11,18 +8,22 @@ import setupSidebarPreload from './preload/sidebar';
 import setupSpellcheckingPreload from './preload/spellchecking';
 import setupTitleChangePreload from './preload/titleChange';
 import setupUserPresencePreload from './preload/userPresence';
+import { setupI18next } from './i18n';
 
+const initialize = async () => {
+	setupErrorHandling('preload');
 
-setupErrorHandling('preload');
-setupContextMenuPreload();
-setupEventsPreload();
-setupJitsiPreload();
-setupLinksPreload();
-setupNotificationsPreload();
-setupSidebarPreload();
-setupSpellcheckingPreload();
-setupTitleChangePreload();
-setupUserPresencePreload();
+	await setupI18next();
 
-window.reloadServer = () => ipcRenderer.sendToHost('reload-server');
-window.i18n = i18n;
+	setupContextMenuPreload();
+	setupEventsPreload();
+	setupJitsiPreload();
+	setupLinksPreload();
+	setupNotificationsPreload();
+	setupSidebarPreload();
+	setupSpellcheckingPreload();
+	setupTitleChangePreload();
+	setupUserPresencePreload();
+};
+
+initialize();