+++ src/preload/userPresence.js
@@ -4,66 +4,42 @@ import { getMeteor, getTracker, getGetUserPreference, getUserPresence } from './
 
 const { powerMonitor } = remote;
 
-const pollUserPresence = (UserPresence, maximumIdleTime) => () => {
-	let isUserPresent = true;
-
-	try {
-		const idleTime = powerMonitor.getSystemIdleTime();
-		isUserPresent = idleTime < maximumIdleTime;
-
-		if (isUserPresent) {
-			UserPresence.setOnline();
-		} else {
-			UserPresence.setAway();
-		}
-	} catch (error) {
-		console.error(error);
-	}
-};
-
 const handleUserPresence = () => {
 	const Meteor = getMeteor();
 	const Tracker = getTracker();
 	const getUserPreference = getGetUserPreference();
 	const UserPresence = getUserPresence();
 
-	if (!Meteor || !Tracker || !getUserPreference || !UserPresence) {
-		return;
-	}
-
-	let intervalID;
-
+	let isAutoAwayEnabled;
+	let idleThreshold;
 	Tracker.autorun(() => {
-		if (intervalID) {
-			clearInterval(intervalID);
-			intervalID = null;
-		}
-
 		const uid = Meteor.userId();
+		isAutoAwayEnabled = getUserPreference(uid, 'enableAutoAway');
+		idleThreshold = getUserPreference(uid, 'idleTimeLimit');
 
-		if (!uid) {
-			return;
+		if (isAutoAwayEnabled) {
+			delete UserPresence.awayTime;
+			UserPresence.start();
 		}
-
-		delete UserPresence.awayTime;
-		UserPresence.awayOnWindowBlur = false;
-		UserPresence.start();
-
-		const isAutoAwayEnabled = getUserPreference(uid, 'enableAutoAway');
-
-		if (!isAutoAwayEnabled) {
-			UserPresence.setOnline();
-			return;
-		}
-
-		const maximumIdleTime = (getUserPreference(uid, 'idleTimeLimit') || 300) * 1000;
-		const idleTimeDetectionInterval = 5000;
-		const callback = pollUserPresence(UserPresence, maximumIdleTime);
-
-		intervalID = setInterval(callback, idleTimeDetectionInterval);
 	});
-};
 
+	let prevState;
+	setInterval(() => {
+		const state = powerMonitor.getSystemIdleState(idleThreshold);
+
+		if (prevState !== state) {
+			const isOnline = !isAutoAwayEnabled || state === 'active' || state === 'unknown';
+
+			if (isOnline) {
+				Meteor.call('UserPresence:online');
+			} else {
+				Meteor.call('UserPresence:away');
+			}
+
+			prevState = state;
+		}
+	}, 1000);
+};
 
 export default () => {
 	window.addEventListener('load', handleUserPresence);