+++ src/scripts/tray.js
@@ -1,34 +1,33 @@
 import { EventEmitter } from 'events';
 
-import { app, Menu, systemPreferences, Tray as TrayIcon } from 'electron';
+import { remote } from 'electron';
+import { t } from 'i18next';
 
-import i18n from '../i18n';
-import { getTrayIconImage } from './icon';
+import { getTrayIconPath } from './icon';
 
+const { app, Menu, nativeTheme, Tray: TrayIcon } = remote;
 
 const getIconTitle = ({ badge }) => (Number.isInteger(badge) ? String(badge) : '');
 
 const getIconTooltip = ({ badge }) => {
-	const appName = app.getName();
-
 	if (badge === '•') {
-		return i18n.__('tray.tooltip.unreadMessage', { appName });
+		return t('tray.tooltip.unreadMessage', { appName: app.name });
 	}
 
 	if (Number.isInteger(badge)) {
-		return i18n.__('tray.tooltip.unreadMention', { appName, count: badge });
+		return t('tray.tooltip.unreadMention', { appName: app.name, count: badge });
 	}
 
-	return i18n.__('tray.tooltip.noUnreadMessage', { appName });
+	return t('tray.tooltip.noUnreadMessage', { appName: app.name });
 };
 
 const createContextMenuTemplate = ({ isMainWindowVisible }, events) => [
 	{
-		label: !isMainWindowVisible ? i18n.__('tray.menu.show') : i18n.__('tray.menu.hide'),
+		label: !isMainWindowVisible ? t('tray.menu.show') : t('tray.menu.hide'),
 		click: () => events.emit('set-main-window-visibility', !isMainWindowVisible),
 	},
 	{
-		label: i18n.__('tray.menu.quit'),
+		label: t('tray.menu.quit'),
 		click: () => events.emit('quit'),
 	},
 ];
@@ -43,10 +42,16 @@ let state = {
 
 const instance = new class Tray extends EventEmitter {}();
 
-let darwinThemeSubscriberId = null;
+const handleThemeUpdate = () => {
+	if (!trayIcon) {
+		return;
+	}
+
+	trayIcon.setImage(getTrayIconPath({ badge: state.badge }));
+};
 
 const createIcon = () => {
-	const image = getTrayIconImage({ badge: state.badge });
+	const image = getTrayIconPath({ badge: state.badge });
 
 	if (trayIcon) {
 		trayIcon.setImage(image);
@@ -56,9 +61,7 @@ const createIcon = () => {
 	trayIcon = new TrayIcon(image);
 
 	if (process.platform === 'darwin') {
-		darwinThemeSubscriberId = systemPreferences.subscribeNotification('AppleInterfaceThemeChangedNotification', () => {
-			trayIcon.setImage(getTrayIconImage({ badge: state.badge }));
-		});
+		nativeTheme.on('updated', handleThemeUpdate);
 	}
 
 	trayIcon.on('click', () => instance.emit('set-main-window-visibility', !state.isMainWindowVisible));
@@ -72,11 +75,9 @@ const destroyIcon = () => {
 		return;
 	}
 
-	if (process.platform === 'darwin' && darwinThemeSubscriberId) {
-		systemPreferences.unsubscribeNotification(darwinThemeSubscriberId);
-		darwinThemeSubscriberId = null;
+	if (process.platform === 'darwin') {
+		nativeTheme.off('updated', handleThemeUpdate);
 	}
-
 
 	trayIcon.destroy();
 	instance.emit('destroyed');